cfsCore/requirements.txt
psycopg2
python-decouple
requests
google-auth
google-api-python-client